Allow default dns servers to be a fallback.
This change set the priority of the DNS servers.

The logic is as follows:

1. If there are manually defined servers use those.  The end user has
manually set what the DNS servers should be and that should be honored.

2. If there are no manually defined DNS servers, use the local host
defined. Of note, Not all os types use NetworkManager and dbus so this
can show up empty

3. If there are no manual or no local host based DNS servers defined
use the default servers.

This will provide that the DNS should always be set to the users
preferences and only include the external cloud based ones if all other
options have been excluded.