Add argument to reboot in fire-and-forget mode
If a reboot is triggered through the HA frontend, we cannot really
send back success information as the HA server is being shutdown during
that processes. In fact, the TCP connection delays shutdown of the core.
Allow to trigger the reboot without waiting for a result.

This allows to prevent an open TCP connection between the frontend and
the HA server. Currently the aiohttp server disconnects after 10s in any
case (and we could shorten that time as well), but rather than shorten
the delay further, just not wait for the operation seems the porper fix.

See also: https://github.com/home-assistant/core/pull/45033