Update default workspace mount directory
Update the workspace directory to match the assumptions made in
`common/rootfs_supervisor/etc/supervisor_scripts/common`  that maps
`/workspaces/supervisor` as the source directory for the supervisor.

This should be a no-op if the name of the git directory is `supervisor` but helps if the user renamed their fork (like I did).