Use aiohttp-fast-url-dispatcher to avoid linear searching to route urls
The default `UrlDispatcher` implementation does a linear search every which can have a significant [TimeComplexity](https://wiki.python.org/moin/TimeComplexity) when dispatching urls when there are a lot of routes. `FastUrlDispatcher` keeps an index of the urls which allows for fast dispatch.

This solution was original implemented in core as
https://github.com/home-assistant/core/blob/241e8560e99da4cbd3c9f763d77cfbd8badb1c66/homeassistant/components/http/__init__.py#L592

It is now in a PyPI package so it can be reused in other projects