Don't remove add-on conifg on add-on removal
Do not remove the add-on config when the add-on gets uninstalled.
The add-on config directory contains user configuration which the
user might like to retain even when uninstalling the add-on. The
user can easily delete the directory on his own since it is user
can access the directory.

In the future we should consider adding an option at uninstall time
to offer removing this directory. But meanwhile not removing it is
the better default.