Add X-Accel-Buffering to disable buffers in proxies
Some proxies, notable NGINX, seem to buffer responses by default. For
the long running log requests this is not helpful: The user won't see
logs since the proxy is buffering the response.

It seems that using http2 or explicit `proxy_buffering: no`` in the
proxy configuration disables the buffering. However, better yet is to
explicitly add the `X-Accel-Buffering: no`` header to the log
responses.