Avoid backup size difference by sorting JSON keys
It seems that the order of keys in the `homeassistant.json` file leads
to a different homeassistant.tar.gz size, which in turn leads to a
different overall backup size. It seems that running the full test
suite compared to run tests individually lead to a different order of
keys in the JSON file.

This PR sorts the keys when dumping the JSON. With that change, the
backup size is consistent.