Fix early Sentry reporting
If an exception occures between when Sentry gets initialized and the
Supervisor state is first set to INITIALIZED (which is only later the
case in `coresys.core.connect()`), the error filtering tries to collect
context which is not ready yet. Make sure that we exit filtering early
just like in the INITIALIZED state.