Use requets http library for connectivity check
The global aiohttp session might cache DNS lookups or pool connections.
This can lead to inaccurate connectivity checks, especially when the
DNS server gets changed after a network setup change.

Instead of using the global aiohttp session for the connectivity checks
this uses the synchronous requests library to perform the connectivity
check. This ensures that the DNS resolution is done fresh for each
check and that the connection is not pooled.

This is another attempt to fix #5332, which got revert in #5851.