Load DNS plug-in early
Load the DNS plug-in early, and create the ClientSession after the
plug-in got loaded and our `/etc/resolv.conf` got written. This makes
sure that Supervisor only reaches out to the internet once we have
DNS setup.

It also side steps an issue with aiodns, which today doesn't reload
`resolv.conf` automatically when it changes. This lead to Supervisor
using the initial `resolv.conf` as created by Docker. It meant that
we did not use the DNS plug-in (and its fallback capabilities) in
Supervisor. Also it meant that changes to the DNS setup at runtime
did not propagate to the aiohttp ClientSession (as observed in #5332).

We could recreate the ClientSession or reload the DNS resolver. But
that seems more complicated. This solution has the added benefit that
requests to internet services as well as our Supervisor connectivity
check is much more deterministic.