Delete Backup files on error
We consider failures during Home Assistatn Core and folder backup as
fatal errors. In this case, we don't return a reference to the backup
file (if run in forground), hence the backup is really considered
fatal. Let's delete the backup files in this case.

Also add unit tests for the backup manager to verify this behavior.