Fix container image detection for aarch64
It seems that on newer Docker (e.g. 28.0.4) the container image built
on the local system on an 64-bit Arm systems returns the architecture
variant just like on 32-bit Arm (but v8 in this case). This is not
expected by the current code. To make Supervisor detect the image
correctly on new and old Docker versions, use substring matching.