Do not backup add-on being uninstalled
When starting a backup we get a object reference to all add-on right
on start. While the backup is running (e.g. the Home Assistant Core
database is being backuped earlier, and can take quite some time),
the user might decide to uninstall an add-on. This will then
result in the add-on metadata being missing by the time the add-on
is being backed up, which will result in an error:

```
KeyError: 'core_assist_microphone'
  File "supervisor/jobs/decorator.py", line 304, in wrapper
    return await self._method(obj, *args, **kwargs)
  File "supervisor/addons/addon.py", line 1339, in backup
    ATTR_USER: self.persist,
  File "supervisor/addons/addon.py", line 267, in persist
    return self.sys_addons.data.user[self.slug]
```

Check the add-on state right before starting the backup to make sure
it is still installed, and if not, report a add-on backup error.